fixed output char

+++ src/octopus/src/serialization/StateDump.cc
@@ -93,7 +93,7 @@ std::string stringState(State const &state_p)
     }
     for(Entity const * ent_l : state_p.getEntities())
     {
-        ss_l << ent_l->_handle.revision << std::endl;
+        ss_l << (int)ent_l->_handle.revision << std::endl;
         ss_l << (ent_l->_alive?1:0) << std::endl;
         ss_l << ent_l->_hp.data() << std::endl;
         ss_l << ent_l->_pos.x.data() << std::endl;
